requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
python-telegram-bot==20.7
python-dotenv==1.0.0
pytz==2023.3
APScheduler==3.10.4
Werkzeug==3.0.1
email-validator==2.1.0
